+++ packages/google_maps_flutter/google_maps_flutter_web/example/integration_test/markers_test.dart
@@ -494,25 +494,28 @@ void main() {
       expect((event as InfoWindowTapEvent).value, equals(const MarkerId('1')));
     });
 
-    // https://github.com/flutter/flutter/issues/80578
     testWidgets('markers with anchor work', (WidgetTester tester) async {
+      const double width = 20;
+      const double height = 30;
+      const Offset defaultOffset = Offset(0.5, 1);
+      const Offset anchorOffset = Offset(1.5, 2);
       final Uint8List bytes = const Base64Decoder().convert(iconImageBase64);
       final Marker marker1 = Marker(
         markerId: const MarkerId('1'),
         icon: BytesMapBitmap(
           bytes,
-          width: 20,
-          height: 30,
+          width: width,
+          height: height,
         ),
       );
       final Marker marker2 = Marker(
         markerId: const MarkerId('2'),
         icon: BytesMapBitmap(
           bytes,
-          width: 20,
-          height: 30,
+          width: width,
+          height: height,
         ),
-        anchor: const Offset(1.5, 2),
+        anchor: anchorOffset,
       );
       final Set<Marker> markers = <Marker>{marker1, marker2};
 
@@ -528,10 +531,10 @@ void main() {
 
       expect(icon1!.anchor, isNotNull);
       expect(icon2!.anchor, isNotNull);
-      expect(icon1.anchor!.x, 20 * 0.5);
-      expect(icon1.anchor!.y, 30 * 1);
-      expect(icon2.anchor!.x, 20 * 1.5);
-      expect(icon2.anchor!.y, 30 * 2);
+      expect(icon1.anchor!.x, width * defaultOffset.dx);
+      expect(icon1.anchor!.y, height * defaultOffset.dy);
+      expect(icon2.anchor!.x, width * anchorOffset.dx);
+      expect(icon2.anchor!.y, height * anchorOffset.dy);
     });
   });
 }