+++ packages/google_maps_flutter/google_maps_flutter_web/CHANGELOG.md
@@ -1,6 +1,10 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
+
 ## 0.5.11
 
-* Adds support for marker anchor
+* Adds support for marker anchor.
 
 ## 0.5.10